Changed 'IsSingleNodeInsideRegion' into extension method
